カメラクラスの作成
